+++ test/unittest/compiler/meson.build
@@ -1,15 +1,22 @@
+test_name = 'unittest_compiler'
+
 test_target = [
-  'unittest_compiler',
-  'unittest_interpreter'
+  'compiler_test_util.cpp',
+  'unittest_compiler.cpp',
+  'unittest_interpreter.cpp',
+  'unittest_realizer.cpp'
 ]
 
-foreach target: test_target
-  exe = executable(
-    target,
-    target + '.cpp',
-    dependencies: nntrainer_test_main_deps,
-    install: get_option('enable-test'),
-    install_dir: application_install_dir
-  )
-  test(target, exe, args: ['--gtest_output=xml:@0@/@1@.xml'.format(meson.build_root(), target)] )
-endforeach
+exe = executable(
+  test_name,
+  test_target,
+  dependencies: [
+    nntrainer_test_main_deps,
+  ],
+  install: get_option('enable-test'),
+  install_dir: application_install_dir
+)
+
+test(test_name, exe,
+  args: '--gtest_output=xml:@0@/@1@.xml'.format(meson.build_root(), test_name)
+)
