+++ test/unittest/compiler/unittest_realizer.cpp
@@ -1,0 +1,15 @@
+// SPDX-License-Identifier: Apache-2.0
+/**
+ * Copyright (C) 2021 Jihoon Lee <jhoon.it.lee@samsung.com>
+ *
+ * @file unittest_realizer.h
+ * @date 09 October 2021
+ * @brief NNTrainer graph realizer related tests
+ * @see	https://github.com/nnstreamer/nntrainer
+ * @author Jihoon Lee <jhoon.it.lee@samsung.com>
+ * @bug No known bugs except for NYI items
+ */
+
+#include <gtest/gtest.h>
+
+TEST(realizer, hello_world) { EXPECT_TRUE(true); }
